refactor: move mcp selector and prompt selector components to libs/shared
