+++ libs/shared/src/components/index.ts
@@ -13,6 +13,8 @@ export * from "./icon-utils";
 export * from "./import-edit-metamodel";
 export * from "./login-page";
 export * from "./login-provider-icon";
+export * from "./mcp";
 export * from "./members";
 export * from "./monaco";
 export * from "./paired-file-upload";
+export * from "./prompts";
