+++ libs/shared/src/components/mcp/index.ts
@@ -1,0 +1,3 @@
+export * from "./mcp-card";
+export * from "./mcp-selector";
+export * from "./mcp-utils";
