+++ libs/shared/src/components/prompts/index.ts
@@ -1,0 +1,1 @@
+export * from "./prompt-selector";
