+++ libs/shared/src/types.ts
@@ -25,7 +25,9 @@ export interface Engine {
 export interface App {
 	project_id: string;
 	project_name: string;
+	project_display_name?: string;
 	description?: string;
+	user_permission?: number;
 }
 
 /**
@@ -307,4 +309,37 @@ export interface User {
 export interface UserAccessRequest {
 	id: string;
 	permission: Role;
+}
+
+export interface MCP {
+	/** Type of the mcp */
+	type: "PROJECT" | "STORAGE" | "DATABASE" | "FUNCTION" | "MODEL" | "VECTOR";
+	/** Id of the mcp */
+	id: string;
+	/** Name of the mcp */
+	name: string;
+	/** Engine subtype (e.g. POSTGRES, OPEN_AI) */
+	subtype?: string;
+	/** Description of the mcp */
+	description?: string;
+	/** Tags of the mcp */
+	tags: string[];
+	permission: "READ_ONLY" | "EDIT" | "OWNER";
+}
+
+export type MCPConfig = Pick<MCP, "type" | "id" | "name"> & {
+	/** Flag to indicate if this MCP comes from a workspace */
+	fromWorkspace?: boolean;
+};
+
+export interface Prompt {
+	id: string;
+	createdBy: string;
+	dateCreated: string;
+	version: number;
+	intent: string;
+	title: string;
+	context: string;
+	tags: string[];
+	global: boolean;
 }
