+++ packages/client/src/utility/index.ts
@@ -1,4 +1,5 @@
 export * from "./general";
+export * from "./mcp-utils";
 export * from "./object";
 export * from "./promise";
 export * from "./tag-color";
