+++ packages/playground/src/components/mcp/index.ts
@@ -1,7 +1,4 @@
 export * from "./agent-selector";
-export * from "./mcp-card";
 export * from "./mcp-overlay";
-export * from "./mcp-selector";
 export * from "./tools-server-view";
 export * from "./tools-view";
-export * from "./utility";
