+++ packages/playground/src/components/mcp/mcp-overlay.tsx
@@ -8,6 +8,7 @@ import type React from "react";
 import { useEffect, useId, useRef, useState } from "react";
 import { useTranslation } from "@semoss/i18n";
 import { usePixel } from "@semoss/sdk/react";
+import { type MCPConfig, MCPSelector } from "@semoss/shared";
 import {
 	Badge,
 	Button,
@@ -23,11 +24,11 @@ import {
 	TabsList,
 	TabsTrigger,
 } from "@semoss/ui/next";
-import type { MCPConfig, Workspace } from "@/types";
+import { useRoot } from "@/hooks";
+import type { Workspace } from "@/types";
+import { mcpToPlatformUrl, splitMcpByType } from "@/utility/mcp-utils";
 import { NewKnowledgeFormBody } from "../knowledge/new-knowledge-form-body";
 import { AgentSelector } from "./agent-selector";
-import { MCPSelector } from "./mcp-selector";
-import { splitMcpByType } from "./utility";
 
 type Tab = "AGENT" | "TOOLBOX" | "KNOWLEDGE";
 type WorkspaceRef = Pick<Workspace, "workspace_id"> &
@@ -79,6 +80,7 @@ export const MCPOverlay: React.FC<MCPOverlayProps> = ({
 	onClose,
 }) => {
 	const { t } = useTranslation(["mcp", "knowledge", "common"]);
+	const { root } = useRoot();
 
 	const [knowledge, setKnowledge] = useState<MCPConfig[]>(
 		() => splitMcpByType(values).knowledge,
@@ -319,6 +321,16 @@ export const MCPOverlay: React.FC<MCPOverlayProps> = ({
 											setView("create")
 										}
 										autoFocus
+										enableKnowledgeMCP={
+											root.theme.featureFlags
+												?.enableKnowledgeMCP
+										}
+										getPlatformUrl={
+											root.theme.featureFlags
+												?.showPlatformLinks
+												? mcpToPlatformUrl
+												: undefined
+										}
 									/>
 								)}
 							</TabsContent>
@@ -332,6 +344,16 @@ export const MCPOverlay: React.FC<MCPOverlayProps> = ({
 										values={toolbox}
 										onChange={setToolbox}
 										autoFocus
+										enableKnowledgeMCP={
+											root.theme.featureFlags
+												?.enableKnowledgeMCP
+										}
+										getPlatformUrl={
+											root.theme.featureFlags
+												?.showPlatformLinks
+												? mcpToPlatformUrl
+												: undefined
+										}
 									/>
 								)}
 							</TabsContent>
