+++ packages/playground/src/components/prompts/index.ts
@@ -1,2 +1,1 @@
 export * from "./prompt-library-dialog";
-export * from "./prompt-selector";
