+++ packages/playground/src/components/room/room-input.tsx
@@ -58,7 +58,6 @@ import {
 import {
 	EnterPlugin,
 	FocusPlugin,
-	isKnowledgeMcp,
 	MCPOverlay,
 	MentionPlugin,
 	PromptLibraryDialog,
@@ -69,6 +68,7 @@ import { RoomInputMenuSlash } from "@/components/room/room-input-menu-slash";
 import { useGracefulErrors, useRoot } from "@/hooks";
 import type { RoomStore } from "@/stores";
 import type { Engine, MCPConfig, Workspace } from "@/types";
+import { isKnowledgeMcp } from "@/utility/mcp-utils";
 import { PromptOptimizer } from "../../components/prompt/PromptOptimizer";
 
 type WorkspaceRef = Pick<Workspace, "workspace_id"> &
