+++ packages/playground/src/types.d.ts
@@ -56,47 +56,8 @@ export interface Instructions {
 	context: string;
 }
 
-export interface MCP {
-	/** Type of the mcp */
-	type: "PROJECT" | "STORAGE" | "DATABASE" | "FUNCTION" | "MODEL" | "VECTOR";
-
-	/** Id of the mcp */
-	id: string;
-
-	/** Name of the mcp */
-	name: string;
-
-	/** Engine subtype (e.g. POSTGRES, OPEN_AI) — used to pick the avatar icon */
-	subtype?: string;
-
-	/** Description of the mcp */
-	description?: string;
-
-	/** Tags of the mcp */
-	tags: string[];
-
-	permission: "READ_ONLY" | "EDIT" | "OWNER";
-}
-
-export type MCPConfig = Pick<MCP, "type" | "id" | "name"> & {
-	/** Flag to indicate if this MCP comes from a workspace */
-	fromWorkspace?: boolean;
-};
-
-/**
- * Item from the prompt library
- */
-export interface Prompt {
-	id: string;
-	createdBy: string;
-	dateCreated: string;
-	version: number;
-	intent: string;
-	title: string;
-	context: string;
-	tags: string[];
-	global: boolean;
-}
+// Re-export types from shared to avoid breaking existing imports
+export type { MCP, MCPConfig, Prompt } from "@semoss/shared";
 
 /**
  * Messages from the backend
